+++ utils/erp.py
@@ -12,6 +12,8 @@ ERP_API_SUBSCRIPTION_KEY = settings.ERP_API_SUBSCRIPTION_KEY
 
 
 def push_fr_data(data, retired=False):
+    if ERP_API_ENDPOINT == 'x':
+        return
     # Contacts
     c_ifrc_names = ",".join(
         data.contacts.filter(ctype__iexact='Federation').values_list('name', flat=True)