+++ CHANGELOG.md
@@ -6,6 +6,32 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## Unreleased
 
+## 1.1.369
+
+### Added
+ - Add global project api endpoints
+
+## 1.1.368
+
+### Added
+ - Handle case where incoming value for personnel deployment from Molnix is null
+
+## 1.1.367
+
+### Added
+ - Return from push_fr_data if ERP_ENDPOINT is 'x' (empty not good: broke tests)
+
+## 1.1.366
+
+### Added
+ - Add country iso and name to district and district centroid tilesets
+ - Bump django from 2.2.22 to 2.2.24
+
+## 1.1.365
+
+### Added
+ - Bump django from 2.2.21 to 2.2.22
+
 ## 1.1.364
 
 ### Added
@@ -1589,7 +1615,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## 0.1.20
 
-[Unreleased]: https://github.com/IFRCGo/go-api/compare/1.1.364...HEAD
+[Unreleased]: https://github.com/IFRCGo/go-api/compare/1.1.369...HEAD
+[1.1.369]: https://github.com/IFRCGo/go-api/compare/1.1.368...1.1.369
+[1.1.368]: https://github.com/IFRCGo/go-api/compare/1.1.367...1.1.368
+[1.1.367]: https://github.com/IFRCGo/go-api/compare/1.1.366...1.1.367
+[1.1.366]: https://github.com/IFRCGo/go-api/compare/1.1.365...1.1.366
+[1.1.365]: https://github.com/IFRCGo/go-api/compare/1.1.364...1.1.365
 [1.1.364]: https://github.com/IFRCGo/go-api/compare/1.1.363...1.1.364
 [1.1.363]: https://github.com/IFRCGo/go-api/compare/1.1.362...1.1.363
 [1.1.362]: https://github.com/IFRCGo/go-api/compare/1.1.361...1.1.362